allow submodule imports from common package
