+++ apps/consent-api/src/routers/status.ts
@@ -19,7 +19,7 @@
 
 import { Router } from 'express';
 // TODO: fix exports from common (e.g. import { APIStatus } from 'common' or 'common/service')
-import { APIStatus } from 'common/src/service/Status';
+import { APIStatus } from 'common/service';
 
 import logger from '@/logger';
