+++ packages/common/src/entity/ConsentCategory.ts
@@ -1,13 +1,5 @@
 import * as z from 'zod';
 
-export const User = z.object({
-	id: z.string(),
-	name: z.string(),
-	email: z.string().email(),
-});
-
-export type User = z.infer<typeof User>;
-
 const CONSENT_CATEGORIES = [
 	'INFORMED_CONSENT',
 	'CONSENT_RELEASE_DATA',
