+++ packages/common/src/entity/User.ts
@@ -1,0 +1,9 @@
+import * as z from 'zod';
+
+export const User = z.object({
+	id: z.string(),
+	name: z.string(),
+	email: z.string().email(),
+});
+
+export type User = z.infer<typeof User>;
