+++ packages/common/src/entity/index.ts
@@ -1,0 +1,2 @@
+export * from './User';
+export * from './ConsentCategory';
