+++ packages/common/src/service/Status.ts
@@ -1,4 +1,4 @@
 export type APIStatus = {
 	status: string;
 	version: string;
-}
+};
